adicionando a view HomeFuncionario

+++ vegalume/Models/Funcionario.cs
@@ -6,7 +6,7 @@
         public int rm { get; set; }
         public string? nome { get; set; }
         public string? senha { get; set; }
-        public decimal? telefone { get; set; }
+        public long? telefone { get; set; }
         public string? email { get; set; }
         public List<Funcionario>? ListaFuncionario { get; set; }
